1、增加了本地文件实时存储的功能； 2、协议对接没有问题，角度按顺序排列版本；

+++ dealusb_msg.h
@@ -26,7 +26,7 @@ public:
 
 
 
-    vector<int> Rece_points;        //保存一帧的点，先角度 后TOF
+    vector<float> Rece_points;        //保存一帧的点，先角度 后TOF
 
     int pointIndex;
